+++ ganttproject/src/main/java/biz/ganttproject/impex/csv/CSVExport.kt
@@ -1,5 +1,5 @@
 /*
-Copyright 2021 Dmitry Barashev, BarD Software s.r.o
+Copyright 2017-2021 Alexander Kurutin, Dmitry Barashev, BarD Software s.r.o
 
 This file is part of GanttProject, an open-source project management tool.
 
@@ -19,12 +19,83 @@ along with GanttProject.  If not, see <http://www.gnu.org/licenses/>.
 package biz.ganttproject.impex.csv
 
 import biz.ganttproject.core.time.GanttCalendar
+import com.google.common.base.Charsets
 import com.google.common.base.Strings
 import net.sourceforge.ganttproject.CustomProperty
 import net.sourceforge.ganttproject.CustomPropertyClass
 import net.sourceforge.ganttproject.CustomPropertyDefinition
+import org.apache.commons.csv.CSVFormat
+import org.apache.commons.csv.CSVPrinter
 import java.io.IOException
+import java.io.OutputStream
+import java.io.OutputStreamWriter
+import java.math.BigDecimal
 
+/**
+ * This is an implementation of SpreadsheetWriter which writes to CSV file
+ *
+ * @author Dmitry Barashev
+ * ------
+ * 2021: translated to Kotlin and significantly updated to allow for data types other than strings
+ *       by Dmitry Barashev
+ * 2017: initially written in Java by Alexander Kurutin.
+ */
+class CsvWriterImpl @JvmOverloads internal constructor(
+  stream: OutputStream, format: CSVFormat, addBom: Boolean = false) : SpreadsheetWriter {
+  private val myCsvPrinter: CSVPrinter
+
+  init {
+    val writer = OutputStreamWriter(stream, Charsets.UTF_8)
+    if (addBom) {
+      writer.write('\ufeff'.toInt())
+    }
+    myCsvPrinter = CSVPrinter(writer, format)
+  }
+
+  @Throws(IOException::class)
+  override fun print(value: String?) {
+    myCsvPrinter.print(value)
+  }
+
+  override fun print(value: Int?) {
+    print(value?.toString())
+  }
+
+  override fun print(value: Double?) {
+    print(value?.toString())
+  }
+
+  override fun print(value: BigDecimal?) {
+    print(value?.toString())
+  }
+
+  override fun print(value: GanttCalendar?) {
+    print(value?.toString())
+  }
+
+  override fun print(value: Boolean?) {
+    print(value?.toString())
+  }
+
+  @Throws(IOException::class)
+  override fun println() {
+    myCsvPrinter.println()
+  }
+
+  @Throws(IOException::class)
+  override fun close() {
+    myCsvPrinter.flush()
+    myCsvPrinter.close()
+  }
+}
+
+/**
+ * Prints the values of the specified custom property definitions using the following rules:
+ * - the order of printing is the same as the order of definitions in the defs list
+ * - if some value is missing, it is replaced with null as String
+ * - the type of the printed value is taken from the definition
+ * - when all definitions are processed, it ends the current record
+ */
 @Throws(IOException::class)
 fun writeCustomPropertyValues(
     writer: SpreadsheetWriter,
